RA-1290 Reverting use of Double to Decimal as certain values in Lats and Longs need higher precisions

+++ src/Esfa.Vacancy.Api.Types/GeoPoint.cs
@@ -5,11 +5,11 @@
         /// <summary>
         /// The longitude.
         /// </summary>
-        public double? Longitude { get; set; }
+        public decimal? Longitude { get; set; }
 
         /// <summary>
         /// The latitude.
         /// </summary>
-        public double? Latitude { get; set; }
+        public decimal? Latitude { get; set; }
     }
 }
